In parallel_test CI job, run 4 proc, 'long' tests first
It is likely that the (Linux) parallel_test job will fail when others
don't if the failure comes from the extra tests due to this job
including the 'long' tests. Switch to running the 'long tests' run
first, so if it is going to fail the job fails as fast as possible.

The long tests were only included in the 2-process run before (because
because Github used to provide 2 CPUs on the CI servers). Move the long
tests to the 4-process run (because Github now provides 4 CPUs on the CI
servers).