Input files for strong and weak scaling performance tests
Designed for ARCHER2 which has 16 cores per NUMA region (and 128 cores
per node, although this is probably only significant for the weak
scaling test that uses 8^n*128 cores).

Note that the runs need to have enough timesteps (1000 seems to be OK)
to give enough accuracy to timings, decreasing the effect of first-step
and any possible compilation time.