Update and optimise r_advection
Remove functions that would 'unnormalise' the shape function for
moment_kinetic runs (which are not yet supported in 2D anyway), in line
with the update in #330 that removed use of the 'unnormalised'
distribution function in z-advection.

Tidy up a bit, e.g. remove an empty loop that probably resulted from a
merge mistake.

Optimise: remove the unnecessary copy into `r.scratch`, which was
previously actually allocating and copying into a temporary array before
copying into `r.scratch`. Instead, just access the view of the buffer
array directly. Also use `r.scratch` for `geofac` and update only in the
z-loop, not inside vperp and vpa loops.