Only call @begin_*_region() if function is doing something
In a couple of places, @begin_*_region() was being called, but the
function might immediately afterwards exit early because it is not
active, or have some conditionals which all happen to be false. In these
cases the @begin_*_region() call is just wasted time, so move it to
where it is definitely needed.