Script to run a 'performance test' run
...where first a single-timestep run is done to make sure everything is
compiled, then the full-length run is done.

So far there is no feature in the `machines` scripts to use this script
for runs, so it is necessary to edit the submission scripts manually.
Could add a flag to `submit-run.sh` and `submit-restart.sh` in the
future to make this more convenient...