Use smoothed upwinding for moment derivatives
In the moment equations, a sharp switch of upwinding does not make sense
in the way that it does for kinetic equations. In the kinetic equations,
particles have a definite direction of motion, so using information only
from where the particle came from makes sense. In the moment equations,
the advection terms are integrals over velocity space, and include
particles moving in both directions. However, we do seem to need
upwinding for numerical stability. So instead do a smooth switch using a
tanh() function so that we upwind when |upar|>~vth, but are centred when
|upar|<<vth.