Split n*upar*d(upar)/dz into upwinded and non-upwinded part
In continuity and pressure equations we have a distinction between
'advective' terms like `upar*d(.)/dz`, which are upwinded, and
'compression' terms like `(.)*d(upar)/dz` which are not upwinded. In the
momentum equation, we previously had `2*n*upar*d(upar)/dz`, which was
upwinded but could be regarded as including both 'advective' and
'compression' terms. When adding ExB drift, to have an 'upwind velocity'
in the momentum equation that is consistent with the continuity and
pressure equations, it is convenient to split
`2*n*upar*d(upar)/dz = n*upar*d(upar)/dz + n*upar*d(upar)/dz` where we
upwind one of the parts (treating as 'advective') and not the other one
(treating as 'compressive').