Only write 'initial_electron' output after the initial solve converges
Now that the kinetic electron solver is reasonably reliable, it is not
critical to be able to restart the electron initialisation. Being able
to skip the initialisation when it has been completed before is now good
enough. Therefore only write to the 'initial_electron' output file once,
when the initial solve has completed.

This change may reduce file sizes a bit, but more importantly will allow
the kinetic electron solve to be run independently at multiple points in
r. Different r-indices may take different timesteps and numbers of steps
to converge, so it would not be straightforward to write intermediate
states of the electron solve from different r-positions to the same
file. For reloading the initial electron solution, it is still
convenient to have a single file (rather than the file-per-r-index that
electron debug I/O will use), so that file must be written just once,
after the solve converges.