Delete unhelpful explicit electron timestep restriction
Limit to fraction of the ion timestep is apparently not required, and
may be unhelpful. When the electron pseudotimestep loop reaches the
maximum iteration count, the ion timestep is reduced and the electron
solve restarted. If reducing the ion timestep restricts the electron
timestep, this can make it harder for the electron solve to converge to
steady state, so the simulation never advances.