For initial electron, save/reload phi, constraint coeffs, fix dt
Boundary values of phi are used as initial guess for iteration when
applying the kinetic electron boundary condition, so save phi to ensure
consistency when reloading from a saved initialisation.

Save the moment constraint coefficients to ensure they are consistent
when restarting. This should not affect simulation results, but not
doing it might cause spurious differences when doing a regression
comparison of reloaded and from-scratch runs.

Fix the 'electron_dt', which should be saved from
`t_params.electron.dt`, not `t_params.electron.dt_before_output`.