Don't save FFTW wisdom when there is no output directory
For runs where we care about performance, there will be an output
directory. For some tests that do not require an output directory, it is
useful to allow chebyshev_pseudospectral coordinates to be created
anyway - in this case do not save wisdom, and use FFTW_ESTIMATE instead
of FFTW_MEASURE to ensure that all processes use the same FFTW
algorithm.