Check arrays only accessed in correct r/z ranges in anysv/anyzv regions
Each subblock of processes in an anysv or anyzv region is supposed to
handle a subset of {r,z} or {r} indices. The process should therefore
not access any indices outside the relevant subset when it is inside an
anysv or anyzv region. Add debug checks activated with the
`@debug_shared_array` macro that verify this.