Replace explicit @boundscheck with @debug_consistency_checks
Now that we do not use `--check-bounds=no`, places where we explicitly
use `@boundscheck` may not be removed in production runs - Julia only
removes `@boundscheck` calls if a function is inlined into an
`@inbounds` block. To ensure that these checks are really removed,
introduce a new debug macro `@debug_consistency_checks` that activates
the checks only if `debug_level` is at least 1.