Add separate third_moment variable in kinetic electron preconditioner
`third_moment` is determined by a constraint equation, as an integral of
the shape function. Using it makes the Jacobian matrix much more sparse,
because there are no (explicit) integral-of-derivative terms - the
integral for third_moment only contributes to the `third_moment` rows of
the matrix (which are small in number as it is a moment variable, not a
shape/distribution function variable), not to the shape function rows.