Represent jacobian_info.matrix in block structured form
The Jacobian matrix for multiple variables is naturally composed of
sub-blocks, where each sub-block is the response of one variable to
a second variable. Store the Jacobian matrix as a Tuple of Tuples of
arrays, where each array is one of these sub-blocks. This slightly
simplifies some indexing, but more importantly gives us the opportunity
to exploit the structure (e.g. known sparsity pattern) of each
sub-block.