Move `preconditioners` setup out of setup_nonlinear_solve()
The exact set of things that need to go into the `preconditioners` Tuple
is problem and solver specific, so makes more sense to define this in
the same place as the preconditioners themselves. Allows the
nonlinear_solvers module to be more generic, which will make it easier
to split into a separate package in future.