Fix application of boundary conditions when there is an r-dimension
When there is an r-dimension, the sheath entrance boundary condition on
the distribution function or shape function depends on the vertical ExB
drift vEz, which depends on Er, which depends on phi being calculated
and having the correct radial boundary conditions applied. It is
therefore not always possible to apply boundary conditions to all ion
quantities at the same time - for standard drift-kinetics, we calculate
moments and phi/Er (and then apply boundary conditions) after applying
the boundary condition to f_i.

Split r boundary conditions into separate functions for moments and f.

Ensure boundary conditions are applied and phi calculated in a
self-consistent order for both moment-kinetic and standard-drift-kinetic
modes.