Update expected results for fokker_planck_time_evolution_tests.jl
Previous expected results for moment-kinetic cases had been run without
enforcing the moment constraints. The option to disable the constraint
forcing was removed so that this is no longer possible (because it is
incorrect, so should not have ever been used).

Also mark a few of the tests as `@long` to reduce normal run-time.