Sensible default for `moments_conservation` setting
Previously `moments_conservation = true` had to be set in the input for
any moment-kinetic run. Now, provide a sensible default (`true` for
moment-kinetic, `false` for standard drift-kinetic) so that the option
does not have to be set explicitly, as the default is the only way it
should ever really be used. Keep the option so that enforcing the
constraints can be disabled explicitly (by setting `moments_conservation
= false`) for testing.