Order 'speed' arrays the same way as distribution functions
The 'speed' arrays previously had their dimensions ordered so that the
dimension in which advection was done with that 'speed' was the
left-most dimension. This was (at least sometimes) an optimisation.
However, now some advection speeds are used to calculate other advection
speeds, so it is not always optimal for the speed arrays to have their
dimensions ordered differently. As the distribution function and its
derivative anyway have to be accessed in a non-contiguous way at the
same time the 'speed' array is accessed, it is not obvious that the
ordering could significantly help vectorisation of loops (etc.?) anyway.

It seems simpler to have all distribution-function-sized arrays have
their dimensions in the same order, so switch to this (will verify that
this change does not lead to significant slow-downs).